requirements.txt
Django==3.2.6
django-filter==2.4.0
django-widget-tweaks==1.4.8
pillow
pytest-django==4.4.0
pytest==6.2.4